Reconnect and rebuild connections when the broker fails. (#3)

+++ java/muon-amqp/src/main/java/io/muoncore/extension/amqp/DefaultAmqpChannel.java
@@ -1,5 +1,6 @@
 package io.muoncore.extension.amqp;
 
+import com.rabbitmq.client.AlreadyClosedException;
 import io.muoncore.Discovery;
 import io.muoncore.channel.impl.StandardAsyncChannel;
 import io.muoncore.channel.support.Scheduler;
@@ -10,6 +11,7 @@ import io.muoncore.message.MuonInboundMessage;
 import io.muoncore.message.MuonMessage;
 import io.muoncore.message.MuonMessageBuilder;
 import io.muoncore.message.MuonOutboundMessage;
+import io.muoncore.transport.TransportEvents;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import reactor.core.Dispatcher;
@@ -22,7 +24,6 @@ import java.util.concurrent.TimeUnit;
 
 public class DefaultAmqpChannel implements AmqpChannel {
 
-    public static final String CHANNEL_SHUTDOWN = "ChannelShutdown";
     private String sendQueue;
     private String receiveQueue;
     private QueueListenerFactory listenerFactory;
@@ -34,6 +35,7 @@ public class DefaultAmqpChannel implements AmqpChannel {
     private Codecs codecs;
     private Discovery discovery;
     private Scheduler scheduler;
+    private boolean channelLive = true;
 
     private CountDownLatch handshakeControl = new CountDownLatch(1);
 
@@ -70,6 +72,11 @@ public class DefaultAmqpChannel implements AmqpChannel {
         sendQueue = serviceName + "-receive-" + UUID.randomUUID().toString();
 
         listener = listenerFactory.listenOnQueue(receiveQueue, msg -> {
+            if (msg == null) {
+                shutdown();
+                return;
+            }
+
             log.trace("Received a message on the receive queue " + msg.getQueueName());
             if ("accepted".equals(msg.getHandshakeMessage())) {
                 log.trace("Handshake completed");
@@ -78,14 +85,14 @@ public class DefaultAmqpChannel implements AmqpChannel {
             }
             if (function != null) {
                 MuonInboundMessage inbound = AmqpMessageTransformers.queueToInbound(msg, codecs);
-                if (inbound.getStep().equals(CHANNEL_SHUTDOWN)) {
+                if (inbound.getStep().equals(TransportEvents.CONNECTION_FAILURE)) {
                     function.apply(null);
                 } else {
                     dispatcher.dispatch(inbound,
                             function::apply, Throwable::printStackTrace);
                 }
             }
-        });
+        }, this::shutdown);
 
         try {
             connection.send(QueueMessageBuilder.queue("service." + serviceName)
@@ -113,6 +120,10 @@ public class DefaultAmqpChannel implements AmqpChannel {
         sendQueue = message.getReplyQueue();
         log.debug("Opening queue to listen " + receiveQueue);
         listener = listenerFactory.listenOnQueue(receiveQueue, msg -> {
+            if (msg == null) {
+                shutdown();
+                return;
+            }
             MuonInboundMessage inboundMessage = AmqpMessageTransformers.queueToInbound(msg, codecs);
             log.debug("Received inbound channel message [" + receiveQueue + "] of type " + message.getProtocol() + ":" + inboundMessage.getStep());
             if (StandardAsyncChannel.echoOut) System.out.println(new Date() + ": Channel[ AMQP Wire >>>>> DefaultAMQPChannel]: Received " + inboundMessage);
@@ -121,7 +132,7 @@ public class DefaultAmqpChannel implements AmqpChannel {
             } else if (function != null) {
                 function.apply(inboundMessage);
             }
-        });
+        }, this::shutdown);
 
         try {
             QueueListener.QueueMessage handshakeResponse = QueueMessageBuilder.queue(message.getReplyQueue())
@@ -139,19 +150,27 @@ public class DefaultAmqpChannel implements AmqpChannel {
 
     @Override
     public void shutdown() {
-
-        try {
-            listener.cancel();
-        } catch (Exception e) {
+        if (channelLive) {
+            log.info("Sending shutdown message to client side channel");
+            function.apply(MuonMessageBuilder.fromService(localServiceName)
+                    .step(TransportEvents.CONNECTION_FAILURE)
+                    .operation(MuonMessage.ChannelOperation.closed)
+                    .contentType("text/plain")
+                    .payload(new byte[0])
+                    .buildInbound());
+//            function.apply(null);
+            channelLive = false;
+            try {
+                listener.cancel();
+                connection.deleteQueue(sendQueue);
+                connection.deleteQueue(receiveQueue);
+            } catch (Throwable e) {
+                log.error("Error cleaning up AMQP Channel {}", e.getMessage());
+            }
         }
-//        if (ownsQueues) {
-            connection.deleteQueue(sendQueue);
-            connection.deleteQueue(receiveQueue);
-//        }
         if (onShutdown != null) {
             this.onShutdown.apply(null);
         }
-
     }
 
     @Override
@@ -161,6 +180,11 @@ public class DefaultAmqpChannel implements AmqpChannel {
 
     @Override
     public void send(MuonOutboundMessage message) {
+        if (!channelLive) {
+            if (StandardAsyncChannel.echoOut) System.out.println(new Date() + ": Channel[ DefaultAMQPChannel ]: CHANNEL CLOSED  " + message);
+            return;
+        }
+
         if (StandardAsyncChannel.echoOut) System.out.println(new Date() + ": Channel[ DefaultAMQPChannel >>>>> AMQP Wire]: Sending " + message);
         if (message != null) {
             log.debug("Sending inbound channel message of type " + message.getProtocol() + "||" + message.getStep());
@@ -170,14 +194,15 @@ public class DefaultAmqpChannel implements AmqpChannel {
                     if (msg.getChannelOperation() == MuonMessage.ChannelOperation.closed) {
                       shutdown();
                     }
-                } catch (IOException e) {
-                    e.printStackTrace();
+                } catch (IOException | AlreadyClosedException e) {
+                    log.warn("Connection Error sending to AMQP Broker, killing channel {}", e.getMessage());
+                    shutdown();
                 }
             }, Throwable::printStackTrace);
         } else {
             send(
                     MuonMessageBuilder.fromService(localServiceName)
-                            .step(CHANNEL_SHUTDOWN)
+                            .step(TransportEvents.CONNECTION_FAILURE)
                             .operation(MuonMessage.ChannelOperation.closed)
                             .contentType("text/plain")
                             .payload(new byte[0])
